Split using a separate archive format instead of a hack
This approach needs a lot more code, but most of it is fairly simple
and the hackiness of the previous solution just didn't cut it.

I can dump my test database using this patch, but there's still a
lot of work to be done; the code is ugly, it doesn't recognize all
object types and no index file is created.  But it works!