Fix schema handling
Previously we were completely incapable of handling objects in
different schemas, so this is a great improvement.

I think it's clearer to always dump the SET statement into the object
file (as opposed to what, for example, pg_restore does); this will
result in more data but then you can quickly see if a function was
moved to a different schema.

Thinking about it now, maybe we should omit that statement for
objects in "public"..