Encode filenames to make sure they work in all filesystems
We use the POSIX's Portable Filename Character Set for this.  Any
character not part of that set will be replaced with an underscore.

Keep in mind that we don't understand (or in some cases even know)
the actual encoding of the data, so the same object name might be
encoded differently on different systems, even if the actual data
isn't any different.